linux-requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-extensions==2.1.4
djangorestframework==3.9.0
docutils==0.14
graphviz==0.10.1
psycopg2==2.7.5
pydot==1.4.0
pydot-ng==2.0.0
pyparsing==2.3.0
pytz==2018.7
six==1.11.0